// src/app/models/File.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface File {
  path: string;
  result_path: string | null;
  relative_path: string;
  sub_directories: Array<string>;
  name: string;
  has_conflicts: boolean;
  has_merged_content: boolean;
}